backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.120.3
h11==0.16.0
httptools==0.7.1
idna==3.11
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.2.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.49.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
transformers>=4.40.0
torch>=2.0.0
accelerate>=1.11.0
safetensors>=0.6.0
scikit-learn>=1.3.0
numpy>=1.24.0
sentencepiece>=0.1.99
protobuf>=4.0.0
google-cloud-translate>=3.15.0
